Initial HTML and JS files for Tableau WDC
The HTML file contains links/includes to various javascript files used by the Web Data Connector infrastructure. The Javascript connects to the purpleair json interface and maps the data into a structure that Tableau can use.  This is based on the USGS example from the WDC community info. (It may have some artifacts from there as a result. It's a very straightforward port, so if anything looks clever assume it wasn't intended that way.)